+++ src/Codeception/Module/REST.php
@@ -357,7 +357,12 @@ class REST extends \Codeception\Module
             $header = str_replace('-','_',strtoupper($header));
             $this->client->setServerParameter("HTTP_$header", $val);
 
-            # Issue #827 - symfony foundation requires 'CONTENT_TYPE' without HTTP_
+            // Issue #1650 - Symfony BrowserKit changes HOST header to request URL
+            if (strtolower($header) == 'host') {
+                $this->client->setServerParameter("HTTP_ HOST", $val);
+            }
+
+            // Issue #827 - symfony foundation requires 'CONTENT_TYPE' without HTTP_
             if ($this->isFunctional and $header == 'CONTENT_TYPE') {
                 $this->client->setServerParameter($header, $val);
             }